fix appdata release version
Dotnet sdk's semantic versioning is not compatible with appdata. Dotnet stable versions are in the form `x.y.z`, while pre-releases are in the form `x.y.z-rc.w`.
eg. `6.0.100` is a stable release, `6.0.100-rc.3` is a pre-release, having them in the correct order in the appdata manifest, with `6.0.100` on top, causes an error